+++ src/app/Component/UtilisateursComponents/utilisateurs-details/utilisateurs-details.component.ts
@@ -6,7 +6,6 @@ import { Observable } from 'rxjs';
 import { Adresse } from 'src/app/Model/Adresse';
 import { AdresseService } from 'src/app/Service/adresse.service';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
-import { InvokeFunctionExpr } from '@angular/compiler';
 
 @Component({
   selector: 'app-utilisateurs-details',
@@ -23,7 +22,8 @@ export class UtilisateursDetailsComponent implements OnInit {
   utilisateurForm: Utilisateur;
   adresseForm: Adresse;
 
-  adresses: Adresse[];
+  allAdresses: Adresse[];
+  existAdresse: Adresse;
 
   constructor(
     private utilisateurService: UtilisateurService,
@@ -55,6 +55,7 @@ export class UtilisateursDetailsComponent implements OnInit {
   remplirAdresseForm() {
     this.adresseForm = new Adresse();
     
+    //Si l'adresse est inchangée
     if(this.adresse.rue.toLocaleLowerCase() == this.editAdresse.get("rue").value.toLocaleLowerCase() &&
     this.adresse.numero == this.editAdresse.get("numero").value &&
     this.adresse.localite.toLocaleLowerCase() == this.editAdresse.get("localite").value.toLocaleLowerCase() &&
@@ -62,24 +63,53 @@ export class UtilisateursDetailsComponent implements OnInit {
     this.adresse.pays.toLocaleLowerCase() == this.editAdresse.get("pays").value.toLocaleLowerCase())
     {
       this.utilisateurForm.adresse = this.utilisateur.adresse;
+      this.modifierUtilisateur();
     }
     else
     {
-      //if(adresse not exist) !!!!!!
+      //Si l'adresse a été modifiée
+      this.getAllAdresses();
+    }
+  }
+
+  adresseExist() {
+    this.existAdresse = new Adresse();
+    this.allAdresses.forEach(elem => {
+      if(elem.rue.toLocaleLowerCase() == this.editAdresse.get("rue").value.toLocaleLowerCase() &&
+      elem.numero == this.editAdresse.get("numero").value &&
+      elem.localite.toLocaleLowerCase() == this.editAdresse.get("localite").value.toLocaleLowerCase() &&
+      elem.codePostal == this.editAdresse.get("codePostal").value &&
+      elem.pays.toLocaleLowerCase() == this.editAdresse.get("pays").value.toLocaleLowerCase())
+        this.existAdresse = elem;
+    });
+
+    //Si l'adresse existe dans la base de données
+    if(this.existAdresse.id == null)
+    {
       this.adresseForm.rue = this.editAdresse.get("rue").value;
       this.adresseForm.numero = this.editAdresse.get("numero").value;
       this.adresseForm.localite = this.editAdresse.get("localite").value;
       this.adresseForm.codePostal = this.editAdresse.get("codePostal").value;
       this.adresseForm.pays = this.editAdresse.get("pays").value;
-
       this.ajouterAdresse();
+    }
+    else
+    {
+      this.utilisateurForm.adresse = this.existAdresse.id;
+      this.adresseForm.rue = this.existAdresse.rue;
+      this.adresseForm.numero = this.existAdresse.numero;
+      this.adresseForm.localite = this.existAdresse.localite;
+      this.adresseForm.codePostal = this.existAdresse.codePostal;
+      this.adresseForm.pays = this.existAdresse.pays;
+      this.modifierUtilisateur();
     }
   }
 
-  adresseExist()
+  getAllAdresses()
   {
     this.adresseService.getAllAdresses().subscribe(data => {
-      this.adresses = data;
+      this.allAdresses = data;
+      this.adresseExist();
     });
   }
 
@@ -97,9 +127,11 @@ export class UtilisateursDetailsComponent implements OnInit {
   }
 
   suppressionUtilisateur(utilisateur: Utilisateur) {
-    this.utilisateurService.supprimerUtilisateur(utilisateur).subscribe(elem => {
-      window.location.reload();
-    }); //ERROR ETC ..... à faire plus tard
+    if(confirm("Voulez-vous supprimer cet utilisateur ? " + utilisateur.nom + " " + utilisateur.prenom)) {
+      this.utilisateurService.supprimerUtilisateur(utilisateur).subscribe(elem => {
+        window.location.reload();
+      }); //ERROR ETC ..... à faire plus tard
+    }
   }
 
 }
